improve error reporting if outer delimiters are available (#194)
Currently, if there's a top-level error in the macro (e.g. a top-level missing
field), then we always get a call_site error. That is because we don't have the
span information corresponding to the outer braces.

This isn't a huge problem if serde_tokenstream is used for the top-level
attribute, but does become one for nested attributes. Trait-based Dropshot
servers are one such case -- `endpoint` annotations are nested under the
top-level and can be pretty far from the call site.

Address this by providing an alternative API that also accepts delimiter spans.

Also remove the `.vscode/settings.json` that's no longer required,
and allow the `target` gitignore rule to match a symlink (all my target
directories are symlinks to a non-btrfs-snapshotted store).